refactor(recipe): create cache keys in cache config file
